+++ RELEASE NOTES.md
@@ -2,6 +2,7 @@ Version 1.6.3 beta
 
 - Added offsetForItemAtIndex: method
 - Fixed Mac OS scrolling glitch when using numberOfVisibleViews
+- iCarousel now respects the layer.doubleSided property of item views
 
 Version 1.6.2